chore: Move components into (nextjs_migration) and remove redundant layout file
