+++ src/app/(nextjs_migration)/components/client/SignInButton.tsx
@@ -5,7 +5,7 @@
 "use client";
 
 import { signIn } from "next-auth/react";
-import { useL10n } from "../../hooks/l10n";
+import { useL10n } from "../../../hooks/l10n";
 
 export const SignInButton = () => {
   const l10n = useL10n();
